requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
langgraph>=0.2.0
anthropic>=0.40.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
